# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal search and RAG tool — indexes data from multiple sources and provides unified search.",
        "title": "Nexus API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/settings/embedding": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns current embedding provider configuration. API keys are masked.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get embedding settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.embeddingSettingsResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the embedding provider. If provider or model changes, automatically triggers a full re-index. Masked API keys (****...) are preserved.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update embedding settings",
                "parameters": [
                    {
                        "description": "Embedding settings",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.embeddingSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.embeddingSettingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/settings/ranking": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the active ranking config — per-source half-life + floor + trust weight, plus the global bonus/trust toggles. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get search ranking tunables",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.rankingSettings"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Persists ranking knobs and hot-swaps them into the in-memory config so the next query sees the new values. Unknown source_type keys in any of the per-source maps are rejected with 400. Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update search ranking tunables",
                "parameters": [
                    {
                        "description": "Ranking settings",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.rankingSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.rankingSettings"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    }
                }
            }
        },
        "/settings/rerank": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns current reranking provider configuration + the min-score floor. API keys are masked.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get rerank settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.rerankSettingsResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the reranking provider + min-score floor. Masked API keys (****...) are preserved. min_score must be in [0,1].",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update rerank settings",
                "parameters": [
                    {
                        "description": "Rerank settings",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.rerankSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.rerankSettingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    }
                }
            }
        },
        "/settings/retention": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the currently-persisted retention policy. Admin only. Reports the hard floor on sweep interval so the UI can surface it as a disabled bound instead of letting the user submit an invalid value.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Get sync-history retention settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.retentionSettingsResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Validates and persists the three retention keys the sweeper reads every tick. Retention values must be non-negative integers (0 disables the rule). The sweep interval is clamped to the hard floor reported by GET.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Update sync-history retention settings",
                "parameters": [
                    {
                        "description": "Retention settings",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.retentionSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.retentionSettingsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    }
                }
            }
        },
        "/settings/retention/sweep": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Triggers a one-shot sweep of the sync_runs history using the currently-persisted retention rules. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settings"
                ],
                "summary": "Run retention cleanup immediately",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.embeddingSettingsRequest": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "ollama_url": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
        "internal_api.embeddingSettingsResponse": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "ollama_url": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.rankingSettings": {
            "type": "object",
            "properties": {
                "known_source_types": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "metadata_bonus_enabled": {
                    "type": "boolean"
                },
                "source_half_life_days": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number"
                    }
                },
                "source_recency_floor": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number"
                    }
                },
                "source_trust_enabled": {
                    "type": "boolean"
                },
                "source_trust_weight": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number"
                    }
                }
            }
        },
...
        "internal_api.rerankSettingsRequest": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "min_score": {
                    "type": "number"
                },
                "model": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
        "internal_api.rerankSettingsResponse": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "min_score": {
                    "description": "MinScore is the post-rerank floor — docs below this are dropped.\nLives on the Rerank endpoint (rather than on Ranking) because it's\nonly meaningful when a reranker is configured.",
                    "type": "number"
                },
                "model": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
        "internal_api.retentionSettingsRequest": {
            "type": "object",
            "properties": {
                "retention_days": {
                    "type": "integer"
                },
                "retention_per_connector": {
                    "type": "integer"
                },
                "sweep_interval_minutes": {
                    "type": "integer"
                }
            }
        },
        "internal_api.retentionSettingsResponse": {
            "type": "object",
            "properties": {
                "min_sweep_interval_minutes": {
                    "type": "integer"
                },
                "retention_days": {
                    "type": "integer"
                },
                "retention_per_connector": {
                    "type": "integer"
                },
                "sweep_interval_minutes": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by either a JWT from /auth/login or a personal access token (nexus_pat_...) from /tokens. The \"Bearer \" prefix is required; it is sent as the Authorization header verbatim.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}